fix(app) #11 change places for the best matching
